Change storage directory structure.
