+++ src/ru/javawebinar/basejava/MainArray.java
@@ -1,7 +1,7 @@
 package ru.javawebinar.basejava;
 
 import ru.javawebinar.basejava.model.Resume;
-import ru.javawebinar.basejava.storage.SortedArrayStorage;
+import ru.javawebinar.basejava.storage.arrayStorage.SortedArrayStorage;
 import ru.javawebinar.basejava.storage.Storage;
 
 import java.io.BufferedReader;
@@ -10,7 +10,7 @@ import java.io.InputStreamReader;
 import java.util.List;
 
 /**
- * Test for ru.javawebinar.basejava.storage.ArrayStorage
+ * Test for ru.javawebinar.basejava.storage.arrayStorage.ArrayStorage
  */
 public class MainArray {
     private final static Storage ARRAY_STORAGE = new SortedArrayStorage();
