+++ src/ru/javawebinar/basejava/storage/mapStorage/MapResumeStorage.java
@@ -1,6 +1,7 @@
-package ru.javawebinar.basejava.storage;
+package ru.javawebinar.basejava.storage.mapStorage;
 
 import ru.javawebinar.basejava.model.Resume;
+import ru.javawebinar.basejava.storage.mapStorage.AbstractMapStorage;
 
 import java.util.HashMap;
