+++ test/ru/javawebinar/basejava/storage/AllStorageTest.java
@@ -2,6 +2,12 @@ package ru.javawebinar.basejava.storage;
 
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
+import ru.javawebinar.basejava.storage.arrayStorage.ArrayStorageTest;
+import ru.javawebinar.basejava.storage.arrayStorage.SortedArrayStorageTest;
+import ru.javawebinar.basejava.storage.listStorage.ListStorageTest;
+import ru.javawebinar.basejava.storage.mapStorage.MapIntegerStorageTest;
+import ru.javawebinar.basejava.storage.mapStorage.MapResumeStorageTest;
+import ru.javawebinar.basejava.storage.mapStorage.MapUuidStorageTest;
 
 /**
  * The class to run all Storage tests.
