--- test/ru/javawebinar/basejava/storage/MapIntegerStorageTest.java
@@ -1,7 +1,0 @@
-package ru.javawebinar.basejava.storage;
-
-public class MapIntegerStorageTest extends AbstractStorageTest{
-    public MapIntegerStorageTest() {
-        super(new MapIntegerStorage());
-    }
-}
