--- test/ru/javawebinar/basejava/storage/MapResumeStorageTest.java
@@ -1,8 +1,0 @@
-package ru.javawebinar.basejava.storage;
-
-public class MapResumeStorageTest extends AbstractStorageTest {
-
-    public MapResumeStorageTest() {
-        super(new MapResumeStorage());
-    }
-}
