--- test/ru/javawebinar/basejava/storage/MapUuidStorageTest.java
@@ -1,7 +1,0 @@
-package ru.javawebinar.basejava.storage;
-
-public class MapUuidStorageTest extends AbstractStorageTest{
-    public MapUuidStorageTest() {
-        super(new MapUuidStorage());
-    }
-}
