+++ test/ru/javawebinar/basejava/storage/listStorage/ListStorageTest.java
@@ -1,4 +1,6 @@
-package ru.javawebinar.basejava.storage;
+package ru.javawebinar.basejava.storage.listStorage;
+
+import ru.javawebinar.basejava.storage.AbstractStorageTest;
 
 public class ListStorageTest extends AbstractStorageTest {
     public ListStorageTest() {
