+++ test/ru/javawebinar/basejava/storage/mapStorage/MapIntegerStorageTest.java
@@ -1,0 +1,10 @@
+package ru.javawebinar.basejava.storage.mapStorage;
+
+import ru.javawebinar.basejava.storage.AbstractStorageTest;
+import ru.javawebinar.basejava.storage.mapStorage.MapIntegerStorage;
+
+public class MapIntegerStorageTest extends AbstractStorageTest {
+    public MapIntegerStorageTest() {
+        super(new MapIntegerStorage());
+    }
+}
