+++ test/ru/javawebinar/basejava/storage/mapStorage/MapResumeStorageTest.java
@@ -1,0 +1,11 @@
+package ru.javawebinar.basejava.storage.mapStorage;
+
+import ru.javawebinar.basejava.storage.AbstractStorageTest;
+import ru.javawebinar.basejava.storage.mapStorage.MapResumeStorage;
+
+public class MapResumeStorageTest extends AbstractStorageTest {
+
+    public MapResumeStorageTest() {
+        super(new MapResumeStorage());
+    }
+}
