+++ test/ru/javawebinar/basejava/storage/mapStorage/MapUuidStorageTest.java
@@ -1,0 +1,10 @@
+package ru.javawebinar.basejava.storage.mapStorage;
+
+import ru.javawebinar.basejava.storage.AbstractStorageTest;
+import ru.javawebinar.basejava.storage.mapStorage.MapUuidStorage;
+
+public class MapUuidStorageTest extends AbstractStorageTest {
+    public MapUuidStorageTest() {
+        super(new MapUuidStorage());
+    }
+}
